# Conflicts:
#	src/main/resources/templates/articleCreateForm.html

+++ src/main/java/dev/sd/project/controllers/ArticleController.java
@@ -5,6 +5,7 @@ import dev.sd.project.model.User;
 import dev.sd.project.repository.ArticleRepository;
 import dev.sd.project.repository.UserRepository;
 import dev.sd.project.service.ArticleService;
+import dev.sd.project.service.UserService;
 import lombok.AllArgsConstructor;
 import lombok.extern.java.Log;
 import org.springframework.stereotype.Controller;
@@ -21,6 +22,7 @@ import java.util.logging.Level;
 @AllArgsConstructor
 public class ArticleController {
     private UserRepository userRepository;
+    private UserService userService;
     private ArticleRepository articleRepository;
     private ArticleService articleService;
 
@@ -45,18 +47,17 @@ public class ArticleController {
     }
 
     @PostMapping("/create")
-    public RedirectView createArticle(String userId,
-                                      String title,
+    public RedirectView createArticle(String title,
+                                      String description,
                                       String content,
-                                      String[] tag )  {
-        User user = userRepository.findByUserId(userId);
+                                      String[] tag)  {
+        User user = userRepository.findByUserId(userService.getCurrentUserId());
 
         if (user == null) {
             log.log(Level.INFO, "User is null");
-
-        }else {
+        } else {
             try {
-                Article article = articleService.createArticle(user,title,content,new HashSet<>(Arrays.asList(tag)));
+                Article article = articleService.createArticle(user,title,description,content,new HashSet<>(Arrays.asList(tag)));
                 return new RedirectView("/article?id="+article.getArticleId());
 
             }catch (Exception e) {}
@@ -65,19 +66,27 @@ public class ArticleController {
         return new RedirectView("/article/create");
 
 }
+
     @GetMapping("/create")
     public ModelAndView createArticleForm(){
         return new ModelAndView("articleCreateForm");
 
     }
+
     @PostMapping ("/edit")
     public RedirectView editArticle(String articleId,
                                     String title,
+                                    String description,
                                     String content,
                                     String[] tag){
+        Article article = articleRepository.findByArticleId(articleId);
+        if (article == null || !userService.getCurrentUserId().equals(article.getWriter().getUserId())){
+            return new RedirectView("/");
+        }
 
         try {
-            Article article = articleService.editArticle(articleId,title,content,new HashSet<>(Arrays.asList(tag)));
+            articleService.editArticle(articleId,title,description,content,new HashSet<>(Arrays.asList(tag)));
+
             return new RedirectView("/article?id="+articleId);
         } catch (Exception e) {
 
@@ -85,13 +94,18 @@ public class ArticleController {
         }
 
     }
+    
     @GetMapping("/edit")
     public ModelAndView editArticleForm(@RequestParam(name = "id")String articleId){
         Article article = articleRepository.findByArticleId(articleId);
         if (article == null){
             return new ModelAndView("articleNotFound");
-
         }
+
+        if (!userService.getCurrentUserId().equals(article.getWriter().getUserId())) {
+            return new ModelAndView("redirect:/");
+        }
+
         ModelAndView articleCreateForm = new ModelAndView("articleCreateForm");
         articleCreateForm.addObject(article);
 
@@ -101,6 +115,10 @@ public class ArticleController {
     }
     @PostMapping("/delete")
     public ModelAndView deleteArticle(String articleId){
+        Article article = articleRepository.findByArticleId(articleId);
+        if (article == null || !userService.getCurrentUserId().equals(article.getWriter().getUserId())){
+            return new ModelAndView("redirect:/");
+        }
 
         try {
             articleService.deleteArticle(articleId);
